requirements.txt
# Payment Data Simulator Dependencies

# Core dependencies
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0
faker>=20.0.0

# Azure SDK
azure-eventhub>=5.11.0
azure-identity>=1.15.0

# Configuration and file watching
watchdog>=3.0.0

# Async support
asyncio-compat>=0.1.0

# Utilities
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Clustering and ML
scikit-learn>=1.3.0
numpy>=1.24.0

